0.6.0-SNAPSHOT Spring Security Configuration
